+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+This file might no be up to date nor complete. Please check the Releases page for more information on versions.
+
+## Version 2024.12.10
+- Fixed offset writing geojson file in segmentation module.
+- Added original_pixelsize parameters to rescale coordinates to match final image size. This allows the use of QuPath pixel classifier trained on resized image, eg. with a lower Resolution parameter (higher pixel size).
+
 ## Version 2024.11.19
 - Initial public release.